Move S3 related classes to the s3 package
